app.js file moved to components
